requirements.txt
appdirs==1.4.3
APScheduler==3.6.1
asn1crypto==0.22.0
awscli==1.14.17
beautifulsoup4==4.5.3
boto3==1.5.7
botocore==1.8.21
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cloudpickle==1.2.1
colorama==0.3.7
configparser==3.7.4
coverage==4.4.1
Django==1.11.22
django-apscheduler==0.3.0
django-nose==1.4.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.4
docutils==0.14
doit==0.29.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
fuzzywuzzy==0.15.0
idna==2.5
ipaddress==1.0.18
jenkinsapi==0.2.12
jmespath==0.9.3
lxml==3.8.0
ntlm-auth==1.0.4
ordereddict==1.1
packaging==16.8
pbr==1.10.0
pprint==0.1
pyasn1==0.4.2
pycparser==2.17
PyJWT==1.7.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-Levenshtein==0.12.0
pytz==2017.2
PyYAML==3.12
requests==2.13.0
requests-ntlm==1.0.0
robobrowser==0.5.3
s3transfer==0.1.12
six==1.11.0
SQLAlchemy==1.3.6
SQLAlchemy-Utils==0.34.1
stevedore==1.17.1
typing==3.5.2.2
tzlocal==2.0.0
urllib3==1.21.1
virtualenv==12.1.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
webencodings==0.5.1
Werkzeug==0.12.1
